movimiento

+++ ms-fyv/src/main/java/com/example/msfyv/repository/FacturaRepository.java
@@ -3,5 +3,5 @@ package com.example.msfyv.repository;
 import com.example.msfyv.entity.Factura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaRepository extends JpaRepository<Factura, Integer> {
+public interface FacturaRepository extends JpaRepository<Factura, Double> {
 }
